[Truffle] Update sprintf per code review comments
